The program compiles once again, but it generates a SIGSEGV when trying to load IT modules. Furthermore GDB dumps core when it tries to analyze the problem. It will be necessary to upgrade to the latest version of GDB to be able to continue this project.
git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@7 88f81a63-fe31-0410-a212-96c38ac4dcd5

+++ cheesetracker/common/components/audio/edit_effect.h
@@ -43,11 +43,11 @@ public:
 
 
 	/* On Select info.. */
-	virtual void selected_notify(Sample_Data *p_data,int p_begin,int p_end)=0;
+	virtual void selected_notify(Sample_Data *p_data,size_t p_begin,size_t p_end)=0;
 	virtual list<Property_Bridge*> get_property_list()=0;
 
 	/* Processing */
-	virtual void process(Sample_Data *p_data,int p_begin,int p_end)=0;
+	virtual void process(Sample_Data *p_data,size_t p_begin,size_t p_end)=0;
 
 
 	/* Info */
